refactor: Move camera command/inquiry/options definitions into a src/camera/ subdirectory.
